requirements.txt

fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2,<3
python-multipart>=0.0.9,<0.1.0


numpy==2.0.2
opencv-python==4.12.0.88
pillow==11.3.0
scikit-learn==1.6.1


faster-whisper==1.0.3


python-dotenv>=1.0,<2.0
requests>=2.31,<3.0


# pydantic-settings>=2.5,<3.0   # Si prefieres cargar .env con Settings
# mediapipe==0.10.14            # SOLO con Python >= 3.10 (en 3.9 no instalar)
